requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django>=3.2.4,<3.3
mysqlclient==2.2.4
pkg_resources==0.0.0
sqlparse==0.4.4
typing_extensions==4.9.0
djangorestframework>=3.12.4,<3.13
python-dotenv>=1.0.1
django-cors-headers>=3.5.0